backend/requirements.txt
# FastAPI Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
neo4j==5.16.0

# AI & NLP
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.20
sentence-transformers==2.3.1
transformers==4.37.2
torch>=2.2.0  # Use latest available version
huggingface-hub==0.20.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
